Extract public methods
