+++ sqle/cmd/scannerd/scanners/common/interface_func.go
@@ -1,0 +1,50 @@
+package common
+
+import (
+	"context"
+	"time"
+	"fmt"
+
+	"github.com/actiontech/sqle/sqle/cmd/scannerd/scanners"
+	"github.com/actiontech/sqle/sqle/pkg/scanner"
+)
+
+func Upload(ctx context.Context, sqls []scanners.SQL, c *scanner.Client, apName string, skipAudit bool) error {
+	// key=fingerPrint val=count
+	counterMap := make(map[string]uint, len(sqls))
+
+	nodeList := make([]scanners.SQL, 0, len(sqls))
+	for _, node := range sqls {
+		counterMap[node.Fingerprint]++
+		if counterMap[node.Fingerprint] <= 1 {
+			nodeList = append(nodeList, node)
+		}
+	}
+
+	reqBody := make([]*scanner.AuditPlanSQLReq, 0, len(nodeList))
+	now := time.Now().Format(time.RFC3339)
+	for _, sql := range nodeList {
+		reqBody = append(reqBody, &scanner.AuditPlanSQLReq{
+			Fingerprint:          sql.Fingerprint,
+			Counter:              fmt.Sprintf("%v", counterMap[sql.Fingerprint]),
+			LastReceiveText:      sql.RawText,
+			LastReceiveTimestamp: now,
+		})
+	}
+
+	err := c.UploadReq(scanner.FullUpload, apName, reqBody)
+	if err != nil {
+		return err
+	}
+
+	if skipAudit {
+		return nil
+	}
+
+	reportID, err := c.TriggerAuditReq(apName)
+	if err != nil {
+		return err
+	}
+	return c.GetAuditReportReq(apName, reportID)
+}
+
