+++ sqle/cmd/scannerd/scanners/mybatis/mybatis.go
@@ -2,20 +2,12 @@ package mybatis
 
 import (
 	"context"
-	"fmt"
-	"io/ioutil"
-	"os"
-	"path"
-	"path/filepath"
-	"strings"
-	"time"
 
 	"github.com/actiontech/sqle/sqle/cmd/scannerd/scanners"
 	driverV2 "github.com/actiontech/sqle/sqle/driver/v2"
 	"github.com/actiontech/sqle/sqle/pkg/scanner"
-	"github.com/actiontech/sqle/sqle/cmd/scannerd/scanners/parse"
+	"github.com/actiontech/sqle/sqle/cmd/scannerd/scanners/common"
 
-	mybatisParser "github.com/actiontech/mybatis-mapper-2-sql"
 	"github.com/sirupsen/logrus"
 )
 
@@ -57,7 +49,7 @@ func New(params *Params, l *logrus.Entry, c *scanner.Client) (*MyBatis, error) {
 }
 
 func (mb *MyBatis) Run(ctx context.Context) error {
-	sqls, err := GetSQLFromPath(mb.xmlDir, mb.skipErrorQuery, mb.skipErrorXml)
+	sqls, err := common.GetSQLFromPath(mb.xmlDir, mb.skipErrorQuery, mb.skipErrorXml, "xml")
 	if err != nil {
 		return err
 	}
@@ -88,103 +80,5 @@ func (mb *MyBatis) SQLs() <-chan scanners.SQL {
 
 func (mb *MyBatis) Upload(ctx context.Context, sqls []scanners.SQL) error {
 	mb.sqls = append(mb.sqls, sqls...)
-
-	// key=fingerPrint val=count
-	counterMap := make(map[string]uint, len(mb.sqls))
-
-	nodeList := make([]scanners.SQL, 0, len(mb.sqls))
-	for _, node := range mb.sqls {
-		counterMap[node.Fingerprint]++
-		if counterMap[node.Fingerprint] <= 1 {
-			nodeList = append(nodeList, node)
-		}
-	}
-
-	reqBody := make([]*scanner.AuditPlanSQLReq, 0, len(nodeList))
-	now := time.Now().Format(time.RFC3339)
-	for _, sql := range nodeList {
-		reqBody = append(reqBody, &scanner.AuditPlanSQLReq{
-			Fingerprint:          sql.Fingerprint,
-			Counter:              fmt.Sprintf("%v", counterMap[sql.Fingerprint]),
-			LastReceiveText:      sql.RawText,
-			LastReceiveTimestamp: now,
-		})
-	}
-
-	err := mb.c.UploadReq(scanner.FullUpload, mb.apName, reqBody)
-	if err != nil {
-		return err
-	}
-
-	if mb.skipAudit {
-		return nil
-	}
-
-	reportID, err := mb.c.TriggerAuditReq(mb.apName)
-	if err != nil {
-		return err
-	}
-	return mb.c.GetAuditReportReq(mb.apName, reportID)
-}
-
-func GetSQLFromPath(pathName string, skipErrorQuery, skipErrorXml bool) (allSQL []driverV2.Node, err error) {
-	if !path.IsAbs(pathName) {
-		pwd, err := os.Getwd()
-		if err != nil {
-			return nil, err
-		}
-		pathName = path.Join(pwd, pathName)
-	}
-
-	fileInfos, err := ioutil.ReadDir(pathName)
-	if err != nil {
-		return nil, err
-	}
-	for _, fi := range fileInfos {
-		var sqlList []driverV2.Node
-		pathJoin := path.Join(pathName, fi.Name())
-
-		if fi.IsDir() {
-			sqlList, err = GetSQLFromPath(pathJoin, skipErrorQuery, skipErrorXml)
-		} else if strings.HasSuffix(fi.Name(), "xml") {
-			sqlList, err = GetSQLFromFile(pathJoin, skipErrorQuery)
-		}
-
-		if err != nil {
-			if skipErrorXml {
-				fmt.Printf("[parse xml file error] parse file %s error: %v\n", pathJoin, err)
-			} else {
-				return nil, fmt.Errorf("parse file %s error: %v", pathJoin, err)
-			}
-		}
-		allSQL = append(allSQL, sqlList...)
-	}
-	return allSQL, err
-}
-
-func GetSQLFromFile(file string, skipErrorQuery bool) (r []driverV2.Node, err error) {
-	content, err := ReadFileContent(file)
-	if err != nil {
-		return nil, err
-	}
-	sqls, err := mybatisParser.ParseXMLQuery(content, skipErrorQuery)
-	if err != nil {
-		return nil, err
-	}
-	for _, sql := range sqls {
-		n, err := parse.Parse(context.TODO(), sql)
-		if err != nil {
-			return nil, err
-		}
-		r = append(r, n...)
-	}
-	return r, nil
-}
-
-func ReadFileContent(file string) (content string, err error) {
-	data, err := ioutil.ReadFile(filepath.Clean(file))
-	if err != nil {
-		return "", err
-	}
-	return string(data), err
+	return common.Upload(ctx, mb.sqls, mb.c, mb.apName, mb.skipAudit)
 }
